+++ assets/javascripts/kitten/karl/meters/karl-counter-block.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react'
+import { CounterBlock } from 'kitten/components/meters/counter-block'
+import { Title } from 'kitten/components/typography/title'
+import { Text } from 'kitten/components/typography/text'
+
+export class KarlCounterBlock extends Component {
+  render() {
+    return (
+      <CounterBlock
+        title="A counter title"
+        subTitle="A counter sub title"
+        counter={13865890}
+      />
+    )
+  }
+}